apps/files/models/__init__.py
from .Config import Config
from .FileAppData import FileAppData

# Disable cache key warnings
import warnings
from django.core.cache import CacheKeyWarning
warnings.simplefilter("ignore", CacheKeyWarning)